refactor: 도메인 개념 분리
