+++ src/main/kotlin/racingcar/domain/Validator.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 import racingcar.util.Constant
 import racingcar.util.Exception
