+++ src/main/kotlin/racingcar/domain/model/RacingGame.kt
@@ -1,4 +1,4 @@
-package racingcar.model
+package racingcar.domain.model
 
 import racingcar.util.Constant
